--- src/main/java/com/testDataModels/Login_credentail.java
@@ -1,14 +1,0 @@
-package com.testDataModels;
-
-import lombok.*;
-
-@ToString
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Login_credentail {
-
-	private String username;
-	private String password;
-}
