+++ src/main/java/com/testdatamodels/LoginCredential.java
@@ -1,0 +1,11 @@
+package com.testdatamodels;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+public class LoginCredential {
+
+	private String username;
+	private String password;
+}
