+++ test-output/Suite/Test.html
@@ -77,11 +77,11 @@ function toggleAllBoxes() {
 <tr>
 <td title='com.runners.TestRunner.runScenario()'><b>runScenario</b><br>Test class: com.runners.TestRunner<br>Test method: Runs Cucumber Scenarios<br>Parameters: "Credit card payment with two item", "Optional[Credit card payment]"</td>
 <td><div><pre>java.lang.ArithmeticException: Person is not eligible to vote
-	at com.stepDef.CheckoutPageSteps.checkout_the_cart(CheckoutPageSteps.java:9)
+	at com.stepdef.CheckoutPageSteps.checkout_the_cart(CheckoutPageSteps.java:9)
 	at ✽.checkout the cart(file:///home/qbuser/Kadavil/EclipseWorkspace/Selenium4CucumberHybridFramework/Features/CreditCardPaymentFeature.feature:26)
 </pre></div><a href='#' onClick='toggleBox("stack-trace746390793", this, "Click to show all stack frames", "Click to hide stack frames")'>Click to show all stack frames</a>
 <div class='stack-trace' id='stack-trace746390793'><pre>java.lang.ArithmeticException: Person is not eligible to vote
-	at com.stepDef.CheckoutPageSteps.checkout_the_cart(CheckoutPageSteps.java:9)
+	at com.stepdef.CheckoutPageSteps.checkout_the_cart(CheckoutPageSteps.java:9)
 	at ✽.checkout the cart(file:///home/qbuser/Kadavil/EclipseWorkspace/Selenium4CucumberHybridFramework/Features/CreditCardPaymentFeature.feature:26)
 </pre></div></td>
 <td>10</td>
