+++ test-output/index.html
@@ -132,31 +132,49 @@
                   <span class="method-name">runScenario</span>
                   <span class="parameters">("Credit card payment with one item", "Optional[Credit card payment]")</span>
                   <div class="stack-trace">org.openqa.selenium.ElementNotInteractableException: element not interactable
-  (Session info: chrome=99.0.4844.51)
-Build info: version: &apos;4.1.1&apos;, revision: &apos;e8fcc2cecf&apos;
-System info: host: &apos;QB-KRT-10370&apos;, ip: &apos;127.0.1.1&apos;, os.name: &apos;Linux&apos;, os.arch: &apos;amd64&apos;, os.version: &apos;5.4.0-104-generic&apos;, java.version: &apos;11.0.14&apos;
-Driver info: org.openqa.selenium.chrome.ChromeDriver
-Command: [8e2e8e30b0bd07a8ed5be441023d4e2f, clickElement {id=910d94a9-a1b0-4894-a736-86e6af67fba8}]
-Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 99.0.4844.51, chrome: {chromedriverVersion: 99.0.4844.51 (d537ec02474b5..., userDataDir: /tmp/.com.google.Chrome.3mQJYQ}, goog:chromeOptions: {debuggerAddress: localhost:46037}, javascriptEnabled: true, networkConnectionEnabled: false, pageLoadStrategy: normal, platform: LINUX, platformName: LINUX, proxy: Proxy(), se:cdp: ws://localhost:46037/devtoo..., se:cdpVersion: 99.0.4844.51, setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:virtualAuthenticators: true}
-Element: [[ChromeDriver: chrome on LINUX (8e2e8e30b0bd07a8ed5be441023d4e2f)] -&gt; xpath: //div[@id=&apos;payment-info-buttons-container&apos;]/button[@class=&apos;button-1 payment-info-next-step-button&apos;]]
-Session ID: 8e2e8e30b0bd07a8ed5be441023d4e2f
-	at java.base/jdk.internal.reflect.NativeConstructorAccessorImpl.newInstance0(Native Method)
-	at java.base/jdk.internal.reflect.NativeConstructorAccessorImpl.newInstance(NativeConstructorAccessorImpl.java:62)
-	at java.base/jdk.internal.reflect.DelegatingConstructorAccessorImpl.newInstance(DelegatingConstructorAccessorImpl.java:45)
-	at java.base/java.lang.reflect.Constructor.newInstance(Constructor.java:490)
-	at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.createException(W3CHttpResponseCodec.java:200)
-	at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:133)
-	at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:53)
-	at org.openqa.selenium.remote.HttpCommandExecutor.execute(HttpCommandExecutor.java:184)
-	at org.openqa.selenium.remote.service.DriverCommandExecutor.invokeExecute(DriverCommandExecutor.java:167)
-	at org.openqa.selenium.remote.service.DriverCommandExecutor.execute(DriverCommandExecutor.java:142)
-	at org.openqa.selenium.remote.RemoteWebDriver.execute(RemoteWebDriver.java:543)
-	at org.openqa.selenium.remote.RemoteWebElement.execute(RemoteWebElement.java:251)
-	at org.openqa.selenium.remote.RemoteWebElement.click(RemoteWebElement.java:77)
-	at com.stepDef.PaymentMethodPageSteps.select_payment_method(PaymentMethodPageSteps.java:24)
-	at ✽.select payment method(file:///home/qbuser/Kadavil/EclipseWorkspace/Selenium4CucumberHybridFramework/Features/CreditCardPaymentFeature.feature:16)
+                      (Session info: chrome=99.0.4844.51)
+                      Build info: version: &apos;4.1.1&apos;, revision: &apos;e8fcc2cecf&apos;
+                      System info: host: &apos;QB-KRT-10370&apos;, ip: &apos;127.0.1.1&apos;, os.name: &apos;Linux&apos;,
+                      os.arch: &apos;amd64&apos;, os.version: &apos;5.4.0-104-generic&apos;, java.version: &apos;11.0.14&apos;
+                      Driver info: org.openqa.selenium.chrome.ChromeDriver
+                      Command: [8e2e8e30b0bd07a8ed5be441023d4e2f, clickElement
+                      {id=910d94a9-a1b0-4894-a736-86e6af67fba8}]
+                      Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 99.0.4844.51,
+                      chrome: {chromedriverVersion: 99.0.4844.51 (d537ec02474b5..., userDataDir:
+                      /tmp/.com.google.Chrome.3mQJYQ}, goog:chromeOptions: {debuggerAddress: localhost:46037},
+                      javascriptEnabled: true, networkConnectionEnabled: false, pageLoadStrategy: normal, platform:
+                      LINUX, platformName: LINUX, proxy: Proxy(), se:cdp: ws://localhost:46037/devtoo..., se:cdpVersion:
+                      99.0.4844.51, setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0,
+                      pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify,
+                      webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true,
+                      webauthn:virtualAuthenticators: true}
+                      Element: [[ChromeDriver: chrome on LINUX (8e2e8e30b0bd07a8ed5be441023d4e2f)] -&gt; xpath:
+                      //div[@id=&apos;payment-info-buttons-container&apos;]/button[@class=&apos;button-1
+                      payment-info-next-step-button&apos;]]
+                      Session ID: 8e2e8e30b0bd07a8ed5be441023d4e2f
+                      at java.base/jdk.internal.reflect.NativeConstructorAccessorImpl.newInstance0(Native Method)
+                      at
+                      java.base/jdk.internal.reflect.NativeConstructorAccessorImpl.newInstance(NativeConstructorAccessorImpl.java:62)
+                      at
+                      java.base/jdk.internal.reflect.DelegatingConstructorAccessorImpl.newInstance(DelegatingConstructorAccessorImpl.java:45)
+                      at java.base/java.lang.reflect.Constructor.newInstance(Constructor.java:490)
+                      at
+                      org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.createException(W3CHttpResponseCodec.java:200)
+                      at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:133)
+                      at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:53)
+                      at org.openqa.selenium.remote.HttpCommandExecutor.execute(HttpCommandExecutor.java:184)
+                      at
+                      org.openqa.selenium.remote.service.DriverCommandExecutor.invokeExecute(DriverCommandExecutor.java:167)
+                      at
+                      org.openqa.selenium.remote.service.DriverCommandExecutor.execute(DriverCommandExecutor.java:142)
+                      at org.openqa.selenium.remote.RemoteWebDriver.execute(RemoteWebDriver.java:543)
+                      at org.openqa.selenium.remote.RemoteWebElement.execute(RemoteWebElement.java:251)
+                      at org.openqa.selenium.remote.RemoteWebElement.click(RemoteWebElement.java:77)
+                      at com.stepdef.PaymentMethodPageSteps.select_payment_method(PaymentMethodPageSteps.java:24)
+                      at ✽.select payment
+                      method(file:///home/qbuser/Kadavil/EclipseWorkspace/Selenium4CucumberHybridFramework/Features/CreditCardPaymentFeature.feature:16)
 
-</div> <!-- stack-trace -->
+                  </div> <!-- stack-trace -->
                   <em>
 (Runs Cucumber Scenarios)                  </em>
                 </div> <!-- method-content -->
